♻️ refactor: Extract common Carousel button and container components
- Extracted common button and container logic into CarouselContainer
- Simplified GridViewContainer and ListViewContainer by using the shared CarouselContainer
