+++ src/components/Media/MediaSection/GridView/index.jsx
@@ -1,33 +1,8 @@
-import styled from '@emotion/styled';
 import { useEffect, useState } from 'react';
 
 import GridView from './GridView';
-import CarouselButton from '../Carousel';
+import Carousel from '../Carousel';
 import useCarousel from '../Carousel/useCarousel';
-
-const MediaContainer = styled.div`
-  position: relative;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const Button = styled(CarouselButton)`
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-
-  ${({ position }) => (position === 'prev' ? 'left: -80px;' : 'right: -80px;')}
-
-  background: none;
-  border: none;
-  padding: 0;
-  margin: 0;
-  cursor: pointer;
-  font: inherit;
-  color: inherit;
-`;
 
 const GridViewContainer = ({ data: pressList }) => {
   const [totalPage, setTotalPage] = useState(0);
@@ -48,11 +23,14 @@ const GridViewContainer = ({ data: pressList }) => {
   }, [pressList]);
 
   return (
-    <MediaContainer>
-      {currentPage > 0 && <Button position="prev" handler={goPrev} />}
+    <Carousel
+      goNext={goNext}
+      goPrev={goPrev}
+      showPrev={currentPage > 0}
+      showNext={currentPage < totalPage}
+    >
       <GridView data={selectedPressData} />
-      {currentPage < totalPage && <Button position="next" handler={goNext} />}
-    </MediaContainer>
+    </Carousel>
   );
 };
