+++ src/components/Media/MediaSection/ListView/index.jsx
@@ -1,35 +1,10 @@
-import styled from '@emotion/styled';
 import { useEffect, useState } from 'react';
 
 import { fieldMap } from '@/utils/constants/constants';
 
-import CarouselButton from '../Carousel';
 import ListView from './ListView';
+import Carousel from '../Carousel';
 import useCarousel from '../Carousel/useCarousel';
-
-const MediaContainer = styled.div`
-  position: relative;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const Button = styled(CarouselButton)`
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-
-  ${({ position }) => (position === 'prev' ? 'left: -80px;' : 'right: -80px;')}
-
-  background: none;
-  border: none;
-  padding: 0;
-  margin: 0;
-  cursor: pointer;
-  font: inherit;
-  color: inherit;
-`;
 
 const ListViewContainer = ({ data: pressList }) => {
   const [totalPage, setTotalPage] = useState(0);
@@ -79,8 +54,7 @@ const ListViewContainer = ({ data: pressList }) => {
   }, [currentPage, totalPage, category]);
 
   return (
-    <MediaContainer>
-      <Button position="prev" handler={goToPrevPage} />
+    <Carousel goPrev={goToPrevPage} goNext={goToNextPage}>
       <ListView
         category={category}
         setCategory={setCategory}
@@ -89,8 +63,7 @@ const ListViewContainer = ({ data: pressList }) => {
         totalPressCount={selectedPressList.length}
         reset={reset}
       />
-      <Button position="next" handler={goToNextPage} />
-    </MediaContainer>
+    </Carousel>
   );
 };
